refactor(fs): decompose oci/unpack.rs into an unpack/ module tree
Splits the ~3650-line OCI layer unpacker into unpack/{mod,entries,
fs_ops,regular_file,device_nodes,whiteout,xattr}.rs by entry kind.
Pure structural move: security allowlists, refusal ladder, and the
unpack_layer signature are byte-identical; every symbol the oci/mod.rs
re-export block names still resolves at the same path.
